refactor(encrypted-archive): move the `EncodeStringFn` type from within `src/core/types/index.ts` to `src/types/builtin.ts`
