+++ packages/encrypted-archive/src/types/builtin.ts
@@ -1,4 +1,13 @@
 /**
+ * Convert a string to an ArrayBufferView or ArrayBuffer.
+ * @see https://nodejs.org/api/buffer.html#static-method-bufferfromstring-encoding
+ * @see https://encoding.spec.whatwg.org/#interface-textencoder
+ * @see https://developer.mozilla.org/docs/Web/API/TextEncoder
+ * @see https://doc.deno.land/deno/stable/~/TextEncoder#encode
+ */
+export type EncodeStringFn = (input: string) => ArrayBufferView | ArrayBufferLike;
+
+/**
  * Takes a value and converts to a human readable string.
  * The string is a single line.
  * This is used to display the value in error messages.
